Factor ObjectMover. Not quite finished, Oracle still needs some attention and there's one more massive method that needs attention.
