+++ relstorage/tests/alltests.py
@@ -21,6 +21,7 @@ import importlib
 
 _include_db = True
 _only_mysql = False
+_only_pgsql = False
 
 def make_suite():
     suite = unittest.TestSuite()
@@ -47,6 +48,8 @@ def make_suite():
         test_modules += db_test_modules
     elif _only_mysql:
         test_modules.extend(x for x in db_test_modules if 'mysql' in x)
+    elif _only_pgsql:
+        test_modules.extend(x for x in db_test_modules if 'postgresql' in x)
 
     for mod_name in test_modules:
         mod = importlib.import_module(mod_name)
@@ -62,6 +65,10 @@ if __name__ == '__main__':
         _include_db = False
         _only_mysql = True
         sys.argv.remove('--only-mysql')
+    if '--only-pgsql' in sys.argv:
+        _include_db = False
+        _only_pgsql = True
+        sys.argv.remove('--only-pgsql')
     logging.basicConfig(level=logging.CRITICAL,
                         format='%(asctime)s %(levelname)-5.5s [%(name)s][%(thread)d:%(process)d][%(threadName)s] %(message)s')
     # We get constant errors about failing to lock a blob file,
