+++ relstorage/tests/testoracle.py
@@ -169,7 +169,7 @@ def test_suite():
         # cx_Oracle blob support can only address up to sys.maxint on
         # 32-bit systems, 4GB otherwise. This takes a great deal of time, however,
         # so allow tuning it down.
-        from relstorage.adapters.mover import ObjectMover
+        from relstorage.adapters.oracle.mover import OracleObjectMover as ObjectMover
         assert hasattr(ObjectMover, 'oracle_blob_chunk_maxsize')
         ObjectMover.oracle_blob_chunk_maxsize = 1024 * 1024 * 100
         large_blob_size = ObjectMover.oracle_blob_chunk_maxsize * 2
